requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# JIRA Integration
jira==3.8.0

# Email Processing
imapclient==3.0.1
email-validator==2.2.0

# Slack Integration
slack-sdk==3.33.4

# NLP and AI (Local LLM)
transformers==4.46.3
torch>=2.0.0
sentence-transformers==3.3.1

# PII/PHI Detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Configuration and Environment
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Vector Database (for knowledge base)
chromadb==0.5.23

# Background Tasks
celery==5.3.6
redis==5.0.1
google-genai>=0.1.0

# Monitoring and Logging
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
unstructured>=0.11.0
